Move parser/decoder to decoder
