+++ plugins/inputs/sflow/decoder/funcs.go
