+++ plugins/inputs/sflow/sflow.go
@@ -13,7 +13,7 @@ import (
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/internal"
 	"github.com/influxdata/telegraf/plugins/inputs"
-	"github.com/influxdata/telegraf/plugins/inputs/sflow/parser/decoder"
+	"github.com/influxdata/telegraf/plugins/inputs/sflow/decoder"
 )
 
 const sampleConfig = `
